+++ examples/dynamic_fanout_container.py
@@ -5,27 +5,25 @@ tasks are also container only.
 More details can be found here: https://github.com/argoproj-labs/hera-workflows/issues/250
 """
 
-from hera import InputFrom, Task, Workflow, WorkflowService
+from hera import Parameter, Task, Workflow
 
-
-with Workflow(
-    "dynamic-fanout-container", service=WorkflowService(host="https://my-argo-server.com", token="my-auth-token"),
-) as w:
-
-    # this can be anything! e.g fetch from some API, then in parallel process all entities; chunk database records
+# assumes you used `hera.set_global_token` and `hera.set_global_host` so that the workflow can be submitted
+with Workflow("dynamic-fanout-container") as w:
+    # this can be anything! e.g. fetch from some API, then in parallel process all entities; chunk database records
     # and process them in parallel, etc.
     generate_task = Task(
-        "generate", image="alpine:latest", command=["echo", '[{"value": "a"}, {"value": "b"}, {"value": "c"}]'],
+        "generate",
+        image="alpine:latest",
+        command=["echo", '[{"value": "a"}, {"value": "b"}, {"value": "c"}]'],
     )
-
     fanout_task = Task(
         "fanout",
-        input_from=InputFrom("generate", ["value"]),
+        with_param=generate_task.get_result(),  # this make the task fan out over the `with_param`
+        # `inputs` sets the correct input parameter so the result is used
+        inputs=[Parameter(name="value", value="{{item.value}}")],
         image="alpine:latest",
-        command=["echo"],
-        args=["{{inputs.parameters.value}}"],
+        command=["echo", "{{inputs.parameters.value}}"],
     )
-
     generate_task >> fanout_task
 
 w.create()