+++ src/hera/io.py
@@ -56,5 +56,11 @@ class IO:
         Validates that the given function and corresponding params fit one another, raises AssertionError if
         conditions are not satisfied.
         """
-        assert len(set([i.name for i in self.inputs])) == len(self.inputs), "input objects must have unique names"
-        assert len(set([o.name for o in self.outputs])) == len(self.outputs), "output objects must have unique names"
+        i_parameters = [obj.as_input() for obj in self.inputs if isinstance(obj, Parameter)]
+        i_artifacts = [obj.as_input() for obj in self.inputs if isinstance(obj, Artifact)]
+        o_parameters = [obj.as_output() for obj in self.outputs if isinstance(obj, Parameter)]
+        o_artifacts = [obj.as_output() for obj in self.outputs if isinstance(obj, Artifact)]
+        assert len(set([i.name for i in i_parameters])) == len(i_parameters), "input parameters must have unique names"
+        assert len(set([i.name for i in i_artifacts])) == len(i_artifacts), "input artifacts must have unique names"
+        assert len(set([o.name for o in o_parameters])) == len(o_parameters), "input parameters must have unique names"
+        assert len(set([o.name for o in o_artifacts])) == len(o_artifacts), "output artifacts must have unique names"