+++ src/hera/task.py
@@ -281,18 +281,41 @@ class Task(IO):
             other.depends += f" {operator} {self.name + condition}"
         return other
 
-    def __rshift__(self, other: "Task") -> "Task":
-        """Sets this task as a dependency of the other passed task.
+    def __rrshift__(self, other: List["Task"]) -> "Task":
+        """Sets the `other` as a dependency of `self`.
+
+        This method piggybacks on the default fallback mechanism of Python, which invokes `__rrshift__` when
+        `__rshift__` fails. Hera uses `__rshift__` (TaskA >> TaskB) for setting task dependency chains. However, that
+        does not natively support lists of items/tasks, so this provides the reverse mechanism - the other here is the
+        list of upstream dependency of `self`.
 
         Parameters
         ----------
-        other: Task
-            The other task(s) to set a dependency for. The new dependency of the task is this task.
+        other: List["Task"]
+            The list of upstream dependencies of this task.
 
         Returns
         -------
         Task
-            The other task that was specified.
+            This task/`self`.
+        """
+        assert isinstance(other, list), f"Unknown type {type(other)} specified using reverse right bitshift operator"
+        for o in other:
+            o.next(self)
+        return self
+
+    def __rshift__(self, other: Union["Task", List["Task"]]) -> Union["Task", List["Task"]]:
+        """Sets this task as a dependency of the other passed task.
+
+        Parameters
+        ----------
+        other: Union["Task", List["Task"]]
+            The other task(s) to set a dependency for. The new dependency of the task is this task.
+
+        Returns
+        -------
+        Union["Task", List["Task"]]
+            The other task/s that was/were specified as the dependencies.
 
         Examples
         --------
@@ -300,9 +323,18 @@ class Task(IO):
         t2 = Task('t2')
         t1 >> t2  # this makes t2 execute AFTER t1
         """
-        return self.next(other)
+        if isinstance(other, Task):
+            return self.next(other)
+        elif isinstance(other, list):
+            for o in other:
+                assert isinstance(
+                    o, Task
+                ), f"Unknown list item type {type(o)} specified using right bitshift operator `>>`"
+                self.next(o)
+            return other
+        raise ValueError(f"Unknown type {type(other)}provided to `__rshift__`")
 
-    def on_workflow_status(self, op: Operator, status: WorkflowStatus) -> "Task":
+    def on_workflow_status(self, status: WorkflowStatus, op: Operator = Operator.Equals) -> "Task":
         """Execute this task conditionally on a workflow status."""
         expression = f"{{{{workflow.status}}}} {op} {status}"
         if self.when:
@@ -329,7 +361,7 @@ class Task(IO):
         other.is_exit_task = True
         return self
 
-    def on_other_result(self, other: "Task", operator: Operator, value: str) -> "Task":
+    def on_other_result(self, other: "Task", value: str, operator: Operator = Operator.Equals) -> "Task":
         """Execute this task based on the `other` result"""
         expression = f"'{other.get_result()}' {operator} {value}"
         if self.when:
@@ -426,42 +458,53 @@ class Task(IO):
             setattr(arguments, "artifacts", artifacts)
         return arguments
 
-    def get_outputs_as(self, name):
+    def get_parameters_as(self, name):
         """Gets all the output parameters from this task"""
         return Parameter(name=name, value=f"{{{{tasks.{self.name}.outputs.parameters}}}}")
 
-    def get_output(
-        self, name: str, path: Optional[str] = None, as_name: Optional[str] = None
-    ) -> Union[Artifact, Parameter]:
-        """Returns an output object in the form of an artifact or parameter based on the name.
+    def get_parameter(self, name: str) -> Parameter:
+        """Returns a Parameter from this tasks' outputs based on the name.
 
         Parameters
         ----------
         name: str
             The name of the parameter to extract as an output.
-        path: Optional[str] = None
-            Path to the file containing the output to share.
-        as_name: Optional[str] = None
-            Name alias for the parameter. This will be used for sharing the output with the consumer.
-
         Returns
         -------
-        Union[Artifact, Parameter]
-            Artifact or Parameter, depending on whatever shareable object has been identified on the task.
+        Parameter
+            Parameter with the same name
+
         """
-        if as_name is None:
-            as_name = name
-        obj = next((output for output in self.outputs if output.name == name), None)
+        parameters = [p for p in self.outputs if isinstance(p, Parameter)]
+        obj = next((output for output in parameters if output.name == name), None)
         if obj:
             if isinstance(obj, Parameter):
                 value = f"{{{{tasks.{self.name}.outputs.parameters.{name}}}}}"
-                return Parameter(as_name, value, default=obj.default)
-            if isinstance(obj, Artifact):
-                if path is None:  # If a new path isn't set, we use the same as the origin
-                    path = obj.path
-                return Artifact(as_name, path=path, from_task=f"{{{{tasks.{self.name}.outputs.artifacts.{name}}}}}")
+                return Parameter(name, value, default=obj.default)
             raise NotImplementedError(type(obj))
-        raise KeyError(f"No output named {name} found")
+        raise KeyError(f"No output parameter named {name} found")
+
+    def get_artifact(self, name: str) -> Artifact:
+        """Returns an Artifact from this tasks' outputs based on the name.
+
+        Parameters
+        ----------
+        name: str
+            The name of the parameter to extract as an output.
+
+        Returns
+        -------
+        Artifact
+            Artifact with the same name
+
+        """
+        artifacts = [p for p in self.outputs if isinstance(p, Artifact)]
+        obj = next((output for output in artifacts if output.name == name), None)
+        if obj:
+            if isinstance(obj, Artifact):
+                return Artifact(name, path=obj.path, from_task=f"{{{{tasks.{self.name}.outputs.artifacts.{name}}}}}")
+            raise NotImplementedError(type(obj))
+        raise KeyError(f"No output artifact named {name} found")
 
     def get_result(self) -> str:
         """Returns the formatted field that points to the result/output of this task"""